requirements.txt
_libgcc_mutex==0.1
_openmp_mutex==5.1
blas==1.0
brotlipy==0.7.0
bzip2==1.0.8
ca-certificates==2023.01.10
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
cryptography==39.0.1
cudatoolkit==11.3.1
cudnn==8.2.1
ffmpeg==4.3
freetype==2.12.1
giflib==5.2.1
gmp==6.2.1
gnutls==3.6.15
h5py==3.7.0
hdf5==1.10.6
idna==3.4
intel-openmp==2023.1.0
joblib==1.1.1
jpeg==9e
lame==3.100
lcms2==2.12
ld_impl_linux-64==2.38
lerc==3.0
libdeflate==1.17
libffi==3.4.4
libgcc-ng==11.2.0
libgfortran-ng==11.2.0
libgfortran5==11.2.0
libgomp==11.2.0
libiconv==1.16
libidn2==2.3.4
libpng==1.6.39
libprotobuf==3.15.8
libstdcxx-ng==11.2.0
libtasn1==4.19.0
libtiff==4.5.0
libunistring==0.9.10
libwebp==1.2.4
libwebp-base==1.2.4
lz4-c==1.9.4
mkl==2023.1.0
mkl-service==2.4.0
mkl_fft==1.3.6
mkl_random==1.2.2
ncurses==6.4
nettle==3.7.3
openh264==2.1.1
openssl==1.1.1t
pillow==9.4.0
pip==23.0.1
pycparser==2.21
pyopenssl==23.0.0
pysocks==1.7.1
python==3.9.16
python_abi==3.9
pytorch==1.12.1
pytorch-mutex==1.0
readline==8.2
requests==2.29.0
scikit-learn==1.2.0
scipy==1.9.3
setuptools==67.8.0
six==1.16.0
sqlite==3.41.2
tbb==2021.8.0
tensorboardx==2.5.1
threadpoolctl==2.2.0
tk==8.6.12
torchaudio==0.12.1
torchvision==0.13.1
typing_extensions==4.5.0
tzdata==2023c
urllib3==1.26.15
wheel==0.38.4
xz==5.4.2
zlib==1.2.13
zstd==1.5.5
absl-py==1.4.0
cachetools==5.3.1
contourpy==1.0.7
cycler==0.11.0
einops==0.6.1
filelock==3.12.0
fonttools==4.39.4
fsspec==2023.5.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
grpcio==1.54.2
huggingface-hub==0.15.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
kiwisolver==1.4.4
loguru==0.7.0
markdown==3.4.3
markupsafe==2.1.2
matplotlib==3.7.1
msgpack==1.0.5
msgpack-numpy==0.4.8
multimethod==1.9.1
numpy==1.23.0
oauthlib==3.2.2
packaging==23.1
protobuf==3.20.1
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pyyaml==5.4.1
pyzmq==25.1.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorpack==0.11
termcolor==2.3.0
timm==0.6.12
toml==0.10.2
torchpack==0.2.2
tqdm==4.65.0
werkzeug==2.3.4
zipp==3.15.0
